+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Do not canonicalize URLs which are proxied.
+
 ## [0.24.0] - 2026-04-19
 
 ### Added